+++ CHANGELOG.md
@@ -69,6 +69,7 @@ VModal
 
 VHalfModal
 
+- Modal height has changed to `0.6`, `0.6`, and `0.9` ratios of screen height as min, ideal, and max heights 
 - Modal how supports landscape sizes
 - Issue with modal snapping to max height if dragged to min when `pullDown` dismiss type is not enabled is fixed